MFB: Fixed compiler warnings
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@277264 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/fileinfo/libmagic/cdf.c
@@ -395,7 +395,7 @@ cdf_read_long_sector_chain(int fd, const cdf_header_t *h, const cdf_sat_t *sat,
 	return 0;
 out:
 	free(scn->sst_tab);
-	return (size_t)-1;
+	return -1;
 }
 
 int
@@ -430,7 +430,7 @@ cdf_read_short_sector_chain(const cdf_header_t *h,
 	return 0;
 out:
 	free(scn->sst_tab);
-	return (size_t)-1;
+	return -1;
 }
 
 int
